release v1.1.0 (#28)
* Update LICENSE

* Adds is_demo to mutations

* Adds demo check for more endpoints

* Adds demo check for project and upload task

* Small adjustments and removes check for demo for create orga

* Adding admin rights for demo version

* Splits app.py into modularized files

* submodule update

* Adds PR changes to work with frontend

* Renames folder handler to middleware

* Submodule update

* Adds restricted endpoint query

* Demo check for rest apis

* Adds blacklist of restricted endpoints

* Adds and slightly changes endpoint

* Removes endpoint notify check for demo

* Changes function name & removes endpoint

* Init sample project state added

* Enum pass without value

* Updates submodule

* Adds import of exported file for knowledge terms

If a file gets imported (for knowledge terms) first the app
will try to read it in the strucutre of exported lists.
If the imported file don't have a fitting structure it gets canceled and the
original import flow will continue

* Adds typing

* Adds yml file

* Add dependabot yml (#20)

* Adds small change

* Adds small change

* Bump pyjwt from 2.1.0 to 2.4.0 (#22)

Bumps [pyjwt](https://github.com/jpadilla/pyjwt) from 2.1.0 to 2.4.0.
- [Release notes](https://github.com/jpadilla/pyjwt/releases)
- [Changelog](https://github.com/jpadilla/pyjwt/blob/master/CHANGELOG.rst)
- [Commits](https://github.com/jpadilla/pyjwt/compare/2.1.0...2.4.0)

---
updated-dependencies:
- dependency-name: pyjwt
  dependency-type: direct:production
...



* fixes #10, adds arm tags to exec envs

* Extend comment

* adds EMBEDDING_CREATION_WARNING to notification data (#24)

* Dependabot pip numpy 1.22.0 (#21)

* Bump numpy from 1.21.0 to 1.22.0

Bumps [numpy](https://github.com/numpy/numpy) from 1.21.0 to 1.22.0.
- [Release notes](https://github.com/numpy/numpy/releases)
- [Changelog](https://github.com/numpy/numpy/blob/main/doc/RELEASE_WALKTHROUGH.rst)
- [Commits](https://github.com/numpy/numpy/compare/v1.21.0...v1.22.0)

---
updated-dependencies:
- dependency-name: numpy
  dependency-type: direct:production
...


* bump further to 1.23.1

* Changes python version to match numpy version

* Changes python version in dockerfile


* Adds qutation escaping (#25)

* Adds qutation escaping

* remove pass

* Update service/search/search.py


* change to union type


* Adds hotfix (#26)

* Adds hotfix

* changes spacy version for drone

* remove specific path

* Removes project export from demo (#27)

+++ graphql_api/mutation/information_source.py
@@ -30,6 +30,7 @@ class CreateInformationSource(graphene.Mutation):
         source_code: Optional[str] = None,
         description: Optional[str] = None,
     ):
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         user = get_user_by_info(info)
         information_source = manager.create_information_source(
@@ -47,6 +48,7 @@ class DeleteInformationSource(graphene.Mutation):
     ok = graphene.Boolean()
 
     def mutate(self, info, project_id: str, information_source_id: str):
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         manager.delete_information_source(project_id, information_source_id)
         notification.send_organization_update(
@@ -63,6 +65,7 @@ class ToggleInformationSource(graphene.Mutation):
     ok = graphene.Boolean()
 
     def mutate(self, info, project_id: str, information_source_id: str):
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         manager.toggle_information_source(project_id, information_source_id)
         notification.send_organization_update(
@@ -79,6 +82,7 @@ class SetAllInformationSourceSelected(graphene.Mutation):
     ok = graphene.Boolean()
 
     def mutate(self, info, project_id: str, value: bool):
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         manager.set_all_information_source_selected(project_id, value)
         notification.send_organization_update(
@@ -108,6 +112,7 @@ class UpdateInformationSource(graphene.Mutation):
         description: Optional[str] = None,
         name: Optional[str] = None,
     ):
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         manager.update_information_source(
             project_id, information_source_id, labeling_task_id, code, description, name
